moved navbar components into a folder and created timeline component
