+++ client/src/components/NavBar/FrontGeneral.css
