+++ client/src/components/NavBar/Timeline.js
@@ -1,0 +1,9 @@
+const Timeline = () => { 
+    return (
+        <div>
+            <p>hi</p>
+        </div>
+    )
+}
+  
+  export default Timeline
